backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
httpx==0.28.1
python-whois==0.9.4
beautifulsoup4==4.12.3
lxml==5.3.0
pydantic==2.10.4
python-dotenv==1.0.1
markdown==3.7
duckduckgo-search==7.5.3
